# Avoid string copies by implementing replace with match